Make render_case_hierarchy no longer a template tag
