+++ corehq/apps/reports/templates/reports/careplan/patient_careplan.html
@@ -25,7 +25,7 @@
                 {# This is needed for rendering the case hierarchy. It's included here to prevent an #}
                 {# import error on the case details page: http://manage.dimagi.com/default.asp?223100 #}
                 {% include 'cloudcare/includes/formplayer-inline.html' %}
-                {% render_case_hierarchy case case_hierarchy_options %}
+                {% include "case/partials/case_hierarchy.html" %}
             </div>
         </div>
     {% endblock patient_content %}
