+++ custom/_legacy/pact/templates/pact/patient/pactpatient_careplan.html
@@ -13,6 +13,6 @@
     {# This is needed for rendering the case hierarchy. It's included here to prevent an #}
     {# import error on the case details page: http://manage.dimagi.com/default.asp?223100 #}
     {% include 'cloudcare/includes/formplayer-inline.html' %}
-    {% render_case_hierarchy case case_hierarchy_options %}
+    {% include "case/partials/case_hierarchy.html" %}
 </div>
 {% endblock %}
